# perf: remove unnecessary deepcopy