--- a/src/poetry/repositories/repository.py
+++ b/src/poetry/repositories/repository.py
@@ -115,6 +115,6 @@ class Repository(AbstractRepository):
         canonicalized_name = canonicalize_name(name)
         for package in self.packages:
             if canonicalized_name == package.name and package.version == version:
-                return package.clone()
+                return package
 
         raise PackageNotFound(f"Package {name} ({version}) not found.")