@radoering 2023-06-09
The more complex the project, the smaller the effect. Tested with a larger project: `17 s` -> `15 s`

Nevertheless, probably significant enough to give up the additional safety net.

> Feedback welcome on how to possible test this 🤷 Although I think the long-run solution would be to refactor the `poetry-core` classes in a more read-only / frozen way

I don't think a sensible test is possible. We won't realize it until we do write at some point.